Move certificate registration and delegation into drep subcommand
