+++ cardano-cli/src/Cardano/CLI/EraBased/Commands/Governance.hs
@@ -9,12 +9,11 @@ module Cardano.CLI.EraBased.Commands.Governance
 
 import           Cardano.Api
 
-import           Cardano.CLI.EraBased.Commands.Governance.Actions (GovernanceActionCmds,
-                   renderGovernanceActionCmds)
+import           Cardano.CLI.EraBased.Commands.Governance.Actions
 import           Cardano.CLI.EraBased.Commands.Governance.Committee
+import           Cardano.CLI.EraBased.Commands.Governance.DRep
 import           Cardano.CLI.EraBased.Vote
 import           Cardano.CLI.Types.Common
-import           Cardano.CLI.Types.Key
 
 import           Data.Text (Text)
 
@@ -32,13 +31,6 @@ data EraBasedGovernanceCmds era
       Lovelace
       (File () Out)
       TransferDirection
-  | EraBasedGovernanceDelegationCertificateCmd
-      StakeIdentifier
-      AnyDelegationTarget
-      (File () Out)
-  | EraBasedGovernanceRegistrationCertificateCmd
-      AnyRegistrationTarget
-      (File () Out)
   | EraBasedGovernanceVoteCmd
       AnyVote
       (File () Out)
@@ -46,10 +38,8 @@ data EraBasedGovernanceCmds era
       (GovernanceCommitteeCmds era)
   | EraBasedGovernanceActionCmds
       (GovernanceActionCmds era)
-  | EraBasedGovernanceDRepGenerateKey
-      (ConwayEraOnwards era)
-      (File (VerificationKey ()) Out)
-      (File (SigningKey ()) Out)
+  | EraBasedGovernanceDRepCmds
+      (GovernanceDRepCmds era)
 
 renderEraBasedGovernanceCmds :: EraBasedGovernanceCmds era -> Text
 renderEraBasedGovernanceCmds = \case
@@ -63,14 +53,11 @@ renderEraBasedGovernanceCmds = \case
     "governance create-mir-certificate transfer-to-treasury"
   EraBasedGovernanceMIRTransfer _ _ _ TransferToReserves ->
     "governance create-mir-certificate transfer-to-reserves"
-  EraBasedGovernanceDelegationCertificateCmd {} ->
-    "governance delegation-certificate"
-  EraBasedGovernanceRegistrationCertificateCmd {} ->
-    "governance registration-certificate"
   EraBasedGovernanceVoteCmd {} ->
     "goverance vote"
   EraBasedGovernanceCommitteeCmds cmds ->
     renderGovernanceCommitteeCmds cmds
   EraBasedGovernanceActionCmds cmds ->
     renderGovernanceActionCmds cmds
-  EraBasedGovernanceDRepGenerateKey{} -> "governance drep key-gen"
+  EraBasedGovernanceDRepCmds cmds ->
+    renderGovernanceDRepCmds cmds
