+++ cardano-cli/src/Cardano/CLI/EraBased/Options/Governance.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE LambdaCase #-}
 
 module Cardano.CLI.EraBased.Options.Governance
   ( EraBasedGovernanceCmds(..)
@@ -17,15 +16,12 @@ import           Cardano.CLI.EraBased.Legacy
 import           Cardano.CLI.EraBased.Options.Common
 import           Cardano.CLI.EraBased.Options.Governance.Actions
 import           Cardano.CLI.EraBased.Options.Governance.Committee
+import           Cardano.CLI.EraBased.Options.Governance.DRep
 import           Cardano.CLI.EraBased.Vote
 import           Cardano.CLI.Types.Common
-import           Cardano.CLI.Types.Key
-import           Cardano.CLI.Types.Legacy
 
 import           Data.Foldable
-import           Data.Functor
 import           Data.Maybe
-import           Data.String
 import           Options.Applicative
 import qualified Options.Applicative as Opt
 
@@ -33,181 +29,11 @@ import qualified Options.Applicative as Opt
 pEraBasedGovernanceCmds :: EnvCli -> CardanoEra era -> Parser (EraBasedGovernanceCmds era)
 pEraBasedGovernanceCmds envCli era =
   asum $ catMaybes
-    [ pEraBasedRegistrationCertificateCmd envCli era
-    , pEraBasedDelegationCertificateCmd envCli era
-    , pEraBasedVoteCmd envCli era
+    [ pEraBasedVoteCmd envCli era
     , pCreateMirCertificatesCmds era
-    , pGovernanceCommitteeCmds era <&> fmap EraBasedGovernanceCommitteeCmds
-    , fmap EraBasedGovernanceActionCmds <$> pGovernanceActionCmds era
-    , pDRepCommands era
-    ]
-
-
--- Registration Certificate related
-
-pEraBasedRegistrationCertificateCmd
-  :: EnvCli -> CardanoEra era -> Maybe (Parser (EraBasedGovernanceCmds era))
-pEraBasedRegistrationCertificateCmd envCli =
-  featureInEra Nothing $ \w ->
-    Just
-      $ subParser "registration-certificate"
-      $ Opt.info (pEraCmd envCli w)
-      $ Opt.progDesc "Create a registration certificate."
- where
-  pEraCmd :: EnvCli -> AnyEraDecider era -> Parser (EraBasedGovernanceCmds era)
-  pEraCmd envCli' = \case
-    AnyEraDeciderShelleyToBabbage sToB ->
-      EraBasedGovernanceRegistrationCertificateCmd
-        <$> asum [ ShelleyToBabbageStakePoolRegTarget sToB
-                     <$> pStakePoolRegistrationParserRequirements envCli'
-                 , ShelleyToBabbageStakeKeyRegTarget sToB
-                     <$> pStakeIdentifier
-                 ]
-        <*> pOutputFile
-
-    AnyEraDeciderConwayOnwards cOn ->
-      EraBasedGovernanceRegistrationCertificateCmd . ConwayOnwardRegTarget cOn
-        <$> asum [ RegisterStakePool cOn
-                     <$> pStakePoolRegistrationParserRequirements envCli'
-                 , RegisterStakeKey cOn
-                     <$> pStakeIdentifier
-                     <*> pKeyRegistDeposit
-                 , RegisterDRep cOn
-                     <$> pDRepVerificationKeyOrHashOrFile
-                     <*> pKeyRegistDeposit
-                 ]
-        <*> pOutputFile
-
---------------------------------------------------------------------------------
-
-data AnyEraDecider era where
-  AnyEraDeciderShelleyToBabbage :: ShelleyToBabbageEra era -> AnyEraDecider era
-  AnyEraDeciderConwayOnwards :: ConwayEraOnwards era -> AnyEraDecider era
-
-instance FeatureInEra AnyEraDecider where
-  featureInEra no yes = \case
-    ByronEra    -> no
-    ShelleyEra  -> yes $ AnyEraDeciderShelleyToBabbage ShelleyToBabbageEraShelley
-    AllegraEra  -> yes $ AnyEraDeciderShelleyToBabbage ShelleyToBabbageEraAllegra
-    MaryEra     -> yes $ AnyEraDeciderShelleyToBabbage ShelleyToBabbageEraMary
-    AlonzoEra   -> yes $ AnyEraDeciderShelleyToBabbage ShelleyToBabbageEraAlonzo
-    BabbageEra  -> yes $ AnyEraDeciderShelleyToBabbage ShelleyToBabbageEraBabbage
-    ConwayEra   -> yes $ AnyEraDeciderConwayOnwards ConwayEraOnwardsConway
-
--- Delegation Certificate related
-
-pEraBasedDelegationCertificateCmd :: EnvCli -> CardanoEra era -> Maybe (Parser (EraBasedGovernanceCmds era))
-pEraBasedDelegationCertificateCmd _envCli =
-  featureInEra Nothing $ \w ->
-    Just
-      $ subParser "delegation-certificate"
-      $ Opt.info (pCmd w)
-      $ Opt.progDesc "Delegation certificate creation."
- where
-  pCmd :: AnyEraDecider era -> Parser (EraBasedGovernanceCmds era)
-  pCmd w =
-    EraBasedGovernanceDelegationCertificateCmd
-      <$> pStakeIdentifier
-      <*> pAnyDelegationCertificateTarget w
-      <*> pOutputFile
-
-  pAnyDelegationCertificateTarget :: ()
-    => AnyEraDecider era
-    -> Parser AnyDelegationTarget
-  pAnyDelegationCertificateTarget e =
-    case e of
-      AnyEraDeciderShelleyToBabbage sbe ->
-        ShelleyToBabbageDelegTarget sbe
-          <$> pStakePoolVerificationKeyOrHashOrFile
-      AnyEraDeciderConwayOnwards cOnwards ->
-        ConwayOnwardDelegTarget cOnwards
-          <$> pStakeTarget cOnwards
-
--- TODO: Conway era AFTER sancho net. We probably want to
--- differentiate between delegating voting stake and reward stake
-pStakeTarget :: ConwayEraOnwards era -> Parser (StakeTarget era)
-pStakeTarget cOnwards =
-  asum
-    [ TargetStakePool cOnwards <$> pStakePoolVerificationKeyOrHashOrFile
-    , TargetVotingDrep cOnwards <$> pDRepVerificationKeyOrHashOrFile
-    , TargetVotingDrepAndStakePool cOnwards
-         <$> pDRepVerificationKeyOrHashOrFile
-         <*> pStakePoolVerificationKeyOrHashOrFile
-    , TargetAlwaysAbstain cOnwards <$ pAlwaysAbstain
-    , TargetAlwaysNoConfidence cOnwards <$ pAlwaysNoConfidence
-   -- TODO: Conway era - necessary constructor not exposed by ledger yet
-   -- so this option is hidden
-    , TargetVotingDRepScriptHash cOnwards <$> pDRepScriptHash
-    ]
-
-pAlwaysAbstain :: Parser ()
-pAlwaysAbstain =
-  flag' () $ mconcat [ long "always-abstain"
-                     , help "Abstain from voting on all proposals."
-                     ]
-
-
-pAlwaysNoConfidence :: Parser ()
-pAlwaysNoConfidence =
-  flag' () $ mconcat [ long "always-no-confidence"
-                     , help "Always vote no confidence"
-                     ]
-
-pDRepScriptHash :: Parser ScriptHash
-pDRepScriptHash =
-  Opt.option scriptHashReader $ mconcat
-    [ Opt.long "drep-script-hash"
-    , Opt.metavar "HASH"
-    , Opt.help $ mconcat
-        [ "DRep script hash (hex-encoded).  "
-        ]
-    , Opt.hidden
-    ]
-
-scriptHashReader :: ReadM ScriptHash
-scriptHashReader = eitherReader $ Right . fromString
-
-pDRepVerificationKeyOrHashOrFile
-  :: Parser (VerificationKeyOrHashOrFile DRepKey)
-pDRepVerificationKeyOrHashOrFile =
-  asum
-    [ VerificationKeyOrFile <$> pDRepVerificationKeyOrFile
-    , VerificationKeyHash <$> pDRepVerificationKeyHash
-    ]
-
-pDRepVerificationKeyHash :: Parser (Hash DRepKey)
-pDRepVerificationKeyHash =
-    Opt.option (pBech32KeyHash AsDRepKey <|> pHexHash AsDRepKey) $ mconcat
-      [ Opt.long "drep-key-hash"
-      , Opt.metavar "HASH"
-      , Opt.help $ mconcat
-          [ "DRep verification key hash (either Bech32-encoded or hex-encoded).  "
-          , "Zero or more occurences of this option is allowed."
-          ]
-      ]
-
-pDRepVerificationKey :: Parser (VerificationKey DRepKey)
-pDRepVerificationKey =
-  Opt.option (readVerificationKey AsDRepKey) $ mconcat
-    [ Opt.long "drep-verification-key"
-    , Opt.metavar "STRING"
-    , Opt.help "DRep verification key (Bech32 or hex-encoded)."
-    ]
-
-pDRepVerificationKeyOrFile :: Parser (VerificationKeyOrFile DRepKey)
-pDRepVerificationKeyOrFile =
-  asum
-    [ VerificationKeyValue <$> pDRepVerificationKey
-    , VerificationKeyFilePath <$> pDRepVerificationKeyFile
-    ]
-
-pDRepVerificationKeyFile :: Parser (VerificationKeyFile In)
-pDRepVerificationKeyFile =
-  fmap File . Opt.strOption $ mconcat
-    [ Opt.long "drep-verification-key-file"
-    , Opt.metavar "FILE"
-    , Opt.help "Filepath of the DRep verification key."
-    , Opt.completer (Opt.bashCompleter "file")
+    , fmap EraBasedGovernanceCommitteeCmds <$> pGovernanceCommitteeCmds era
+    , fmap EraBasedGovernanceActionCmds    <$> pGovernanceActionCmds era
+    , fmap EraBasedGovernanceDRepCmds      <$> pGovernanceDRepCmds envCli era
     ]
 
 --------------------------------------------------------------------------------
@@ -243,10 +69,7 @@ pAnyVotingStakeVerificationKeyOrHashOrFile =
        , AnyStakePoolVerificationKeyOrHashOrFile <$> pStakePoolVerificationKeyOrHashOrFile
        ]
 
-
-
 --------------------------------------------------------------------------------
-
 
 pCreateMirCertificatesCmds :: CardanoEra era -> Maybe (Parser (EraBasedGovernanceCmds era))
 pCreateMirCertificatesCmds =
@@ -300,19 +123,3 @@ pMIRTransferToReserves w =
     <*> pOutputFile
     <*> pure TransferToReserves
 
---------------------------------------------------------------------------------
-
-pDRepCommands :: ()
-  => CardanoEra era
-  -> Maybe (Parser (EraBasedGovernanceCmds era))
-pDRepCommands era = do
-  w <- maybeFeatureInEra era
-  pure $
-    subParser "drep"
-    $ Opt.info (pKeyGen w)
-    $ Opt.progDesc "Delegate Representative commands."
-  where
-    pKeyGen w =
-      subParser "key-gen"
-      $ Opt.info (EraBasedGovernanceDRepGenerateKey w <$> pVerificationKeyFileOut <*> pSigningKeyFileOut)
-      $ Opt.progDesc "Generate Delegate Representative verification and signing keys."
