+++ cardano-cli/test/cardano-cli-golden/files/golden/help/experimental_governance.cli
@@ -1,11 +1,4 @@
-Usage: cardano-cli experimental governance 
-                                             ( registration-certificate
-                                             | delegation-certificate
-                                             | vote
-                                             | committee
-                                             | action
-                                             | drep
-                                             )
+Usage: cardano-cli experimental governance (vote | committee | action | drep)
 
   Era-based governance commands
 
@@ -13,8 +6,6 @@ Available options:
   -h,--help                Show this help text
 
 Available commands:
-  registration-certificate Create a registration certificate.
-  delegation-certificate   Delegation certificate creation.
   vote                     Vote creation.
   committee                Committee member commands.
   action                   Governance action commands.
