--- cardano-cli/test/cardano-cli-golden/files/golden/help/shelley_governance_delegation-certificate.cli
@@ -1,31 +1,0 @@
-Usage: cardano-cli shelley governance delegation-certificate 
-                                                               ( --stake-verification-key STRING
-                                                               | --stake-verification-key-file FILE
-                                                               | --stake-script-file FILE
-                                                               | --stake-address ADDRESS
-                                                               )
-                                                               ( --stake-pool-verification-key STRING
-                                                               | --cold-verification-key-file FILE
-                                                               | --stake-pool-id STAKE_POOL_ID
-                                                               )
-                                                               --out-file FILE
-
-  Delegation certificate creation.
-
-Available options:
-  --stake-verification-key STRING
-                           Stake verification key (Bech32 or hex-encoded).
-  --stake-verification-key-file FILE
-                           Filepath of the staking verification key.
-  --stake-script-file FILE Filepath of the staking script.
-  --stake-address ADDRESS  Target stake address (bech32 format).
-  --stake-pool-verification-key STRING
-                           Stake pool verification key (Bech32 or hex-encoded).
-  --cold-verification-key-file FILE
-                           Filepath of the stake pool verification key.
-  --stake-pool-id STAKE_POOL_ID
-                           Stake pool ID/verification key hash (either
-                           Bech32-encoded or hex-encoded). Zero or more
-                           occurences of this option is allowed.
-  --out-file FILE          The output file.
-  -h,--help                Show this help text
